pyproject.toml
[tool.poetry]
name = "discordpy-logging-handler"
version = "0.1.0"
description = "Forward Discord bot logs to Discord Text channel."
authors = ["zztkm <takumi.basket1682@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = ""

include = [
  "docs",
  "poetry.lock",
  "tests",
  "tox.ini",
]

exclude = [
  "docs/_build",
]

[tool.poetry.dependencies]
python = "^3.6.2"
Sphinx = {version = "^4.3.2", optional = true, extras = ["docs"]}
sphinx-autodoc-typehints = {version = "1.12.0", optional = true, extras = ["docs"]}
sphinx-rtd-theme = {version = "^1.0.0", optional = true, extras = ["docs"]}

[tool.poetry.dev-dependencies]
"discord.py" = "^1.7.3"
pytest = "^6.2.5"
tox = "^3.24.5"
pre-commit = "^2.16.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
